requirements.txt
Django==1.7
django-embed-video==0.11
django-taggit==0.13.0
MySQL-python==1.2.5
pbr==0.10.8
requests==2.6.0
six==1.9.0
stevedore==1.2.0
virtualenv==12.0.7
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.2